server/requirements.txt
google-cloud-firestore==2.3.4